sched/fair: Remove and disable LB_BIAS feature
This is based on a recent patch commit:
https://lore.kernel.org/patchwork/patch/973154/

Here it shows that LB_BIAS is being neglected by the scheduler wherein such feature is simply not being used and only serves as dead weight to the code. Thus, let's nuke it alongside the load_idx code that seems to be no longer in use after removing the load balancing bias.

Signed-off-by: ederekun <sedrickvince@gmail.com>